# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Habbis mobile API",
        "description": "This is APIs for habbis social mobile app",
        "termsOfService": "This API document is belong to habbis mobile team.",
        "contact": {
            "email": "songxunzhao1991@gmail.com"
        },
        "license": {
            "name": "Private License",
            "url": ""
        },
        "version": "1.0.0"
    },
    "host": "localhost",
    "basePath": "/habbis/public/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "api/event": {
            "get": {
                "tags": [
                    "Event"
                ],
                "summary": "Get event list",
                "description": "Get event list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "properties": {
                                "code": {
                                    "description": "Response code",
                                    "type": "integer",
                                    "default": 200
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Event"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Event"
                ],
                "summary": "Create event",
                "description": "Create event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "description": "Register data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Event"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "properties": {
                                "code": {
                                    "description": "Response code",
                                    "type": "integer",
                                    "default": 200
                                },
                                "data": {
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "api/event/{event_id}/attend": {
            "get": {
                "tags": [
                    "Event"
                ],
                "summary": "Attend event",
                "description": "Attend event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "path",
                        "description": "Event id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "properties": {
                                "code": {
                                    "description": "Response code",
                                    "type": "integer",
                                    "default": 200
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "api/event/{event_id}": {
            "get": {
                "tags": [
                    "Event"
                ],
                "summary": "Get event detail",
                "description": "Get event detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "path",
                        "description": "Event id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "properties": {
                                "code": {
                                    "description": "Response code",
                                    "type": "integer",
                                    "default": 200
                                },
                                "data": {}
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "api/user/invite": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Invite user by email",
                "description": "Invite user by email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "description": "Register data",
                        "required": true,
                        "schema": {
                            "required": [
                                "email"
                            ],
                            "properties": {
                                "email": {
                                    "description": "Invitee's email",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "properties": {
                                "code": {
                                    "description": "Response code",
                                    "type": "integer",
                                    "default": 200
                                },
                                "message": {
                                    "description": "Response message",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "api/user/profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Profile",
                "description": "Get user profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "properties": {
                                "code": {
                                    "description": "Response code",
                                    "type": "integer",
                                    "default": 200
                                },
                                "data": {
                                    "description": "Response data",
                                    "$ref": "#/definitions/User"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Profile",
                "description": "Get user profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "properties": {
                                "code": {
                                    "description": "Response code",
                                    "type": "integer",
                                    "default": 200
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Event": {
            "properties": {
                "img": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "from": {
                    "type": "string",
                    "format": "date-time"
                },
                "to": {
                    "type": "string",
                    "format": "date-time"
                },
                "description": {
                    "type": "string"
                },
                "venue": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time"
                }
            },
            "xml": {
                "name": "Event"
            }
        },
        "User": {
            "properties": {
                "name": {
                    "description": "",
                    "type": "string"
                },
                "age": {
                    "description": "",
                    "type": "integer"
                },
                "location": {
                    "description": "",
                    "type": "string"
                },
                "lat": {
                    "description": "",
                    "type": "number",
                    "format": "float"
                },
                "lng": {
                    "description": "",
                    "type": "number",
                    "format": "float"
                },
                "job_name": {
                    "description": "",
                    "type": "string"
                },
                "profile_img": {
                    "description": "",
                    "type": "string"
                }
            },
            "xml": {
                "name": "User"
            }
        },
...
    }
}